fb/ram/gf10x: parse memory clock/tweak tables up-front
Ported from ram/gk104, and will be used for similar reasons (noted in
the code's comments).

This is a little different from gk104's implementation, in that it
completely ignores rammap entries that have min/max clock set to 0.

We probably want to check if we need the same in gk104.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>