core: implement a base abstraction for script generation
We have a number of different places where we need to generate "scripts"
of register poking to be executed at a later time (or from a coprocessor
on the GPU).

There are currently multiple implementations that duplicate the majority
of the state-tracking logic that this requires.

In addition to making that logic common, this abstraction also:

- Removes the need to explicitly define structures to track state for
  every register that one might want to access from the script.

- Allows implementations to share script generation code, which could
  not be done cleanly before due to the above.

- Makes it possible to use existing subdev functions during script
  generation, rather than re-inventing the wheel (ie. mclk GPIOs).

- Allows allows us to support NVIDIA's PMU SEQ script interface with
  our existing mclk code, *without* having to modify our FW to match.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>