TBDmmu/gp100-: add privileged methods for fault replay/cancel
Host methods exist to do at least some of what we need, but we are not
currently pushing replay/cancels through a channel like UVM does as it's
not clear whether it's necessary in our case (UVM also updates PTEs with
the GPU).

UVM pushes a software method for fault cancels on Pascal, which could
indicate that we potentially need some special SW handling anyway.

Keep it simple for now, until we figure out exactly what we need here.

TBD: not sure targeted fault cancel is sufficient

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>